+++ R/progress.R
@@ -164,8 +164,10 @@ progress_bar <- R6Class("progress_bar",
     },
     tick = function(len = 1, tokens = list()) {
       pb_tick(self, private, len, tokens) },
-    update = function(ratio, tokens = list()) { 
-      pb_update(self, private, ratio, tokens) }
+    update = function(ratio, tokens = list()) {
+      pb_update(self, private, ratio, tokens) },
+    message = function(msg) {
+      pb_message(self, private, msg) }
   ),
 
   private = list(
@@ -187,6 +189,7 @@ progress_bar <- R6Class("progress_bar",
     ),
     callback = NULL,
     clear = NULL,
+    finished = FALSE,
     show_after = NULL,
     last_draw = "",
 
@@ -248,6 +251,7 @@ pb_tick <- function(self, private, len, tokens) {
 
   assert_scalar(len)
   assert_named_or_empty_list(tokens)
+  stopifnot(!private$finished)
 
   if (private$first) {
     private$first <- FALSE
@@ -393,11 +397,30 @@ pb_render <- function(self, private, tokens) {
 
 pb_update <- function(self, private, ratio, tokens) {
   assert_ratio(ratio)
+  stopifnot(!private$finished)
+
   goal <- floor(ratio * private$total)
   self$tick(goal - private$current, tokens)
 }
 
+pb_message <- function(self, private, msg) {
+  assert_character(msg)
+  stopifnot(!private$finished)
+
+  if (!private$supported) {
+    cat(msg, sep = "\n", file = private$stream)
+  } else {
+    clear_line(private$stream, private$width)
+    cursor_to_start(private$stream)
+    cat(msg, sep = "\n", file = private$stream)
+    if (!private$finished) {
+      cat(private$last_draw, file = private$stream)
+    }
+  }
+}
+
 pb_terminate <- function(self, private) {
+  private$finished <- TRUE
   if (!private$supported || !private$toupdate) return(invisible())
   if (private$clear) {
     clear_line(private$stream, private$width)